Easily and rapidly generate raster image data with support for Plotly.js

Create a rasterly object, to which aggregation layers may be added. This function is the first step in the process to generate raster image data using the rasterly package.

rasterly(data = NULL, mapping = aes(), ..., plot_width = 600,
  plot_height = 600, x_range = NULL, y_range = NULL,
  background = "white", color = NULL, show_raster = TRUE,
  drop_data = FALSE, variable_check = FALSE)

Arguments

data
Dataset to use for generating the plot. If not provided, data must be supplied in each layer of the plot. For best performance, particularly when processing large datasets, use of data.table is recommended.

mapping
Default list of aesthetic mappings to use for plot. The same with ggplot2 aes. See details.

...
Other arguments which will be passed through to layers.

plot_width
Integer. The width of the image to plot; must be a positive integer. A higher value indicates a higher resolution.

plot_height
Integer. The height of the image to plot; must be a positive integer. A higher value indicates a higher resolution.

x_range
Vector of type numeric. The range of x; it can be used to clip the image. For larger datasets, providing x_range may result in improved performance.

y_range
Vector of type numeric. The range of y; it can be used to clip the image. For larger datasets, providing y_range may result in improved performance.

background
Character. The background color of the image to plot.

color
Vector of type character. It will determine this color vector is a color_map or color_key automatically.

color_map: It has Color(s) used to draw each pixel. The color_map is extended by linear interpolation independently for RGB. The darkness of the mapped color depends upon the values of the aggregation matrix.

color_key: Vector of type character. The color_key is used for categorical variables; it is passed when the color aesthetic is provided.

show_raster
Logical. Should the raster be displayed?

drop_data
Logical. When working with large datasets, drops the original data once processed according to the provided aes() parameters, using the remove() function. See details for additional information.

variable_check
Logical. If TRUE, drops unused columns to save memory; may result in reduced performance.

Value

An environment wrapped by a list which defines the properties of the raster data to be generated.

Details

The rasterly package currently supports five aesthetics via aes(): x, y, on, color, and size. The "on" aesthetic specifies the variable upon which the reduction function should be applied to generate the raster data.

drop_data can help save space, particularly when large datasets are used. However, dropping the original dataset may result in errors when attempting to set or update aes() parameters within rasterly layers.

Note

Calling rasterly() without providing rasterly_...() layers has no effect. More info can be found in README.md

See also

Examples

if (FALSE) {
 if(requireNamespace("data.table")) {
 url1 <- "https://raw.githubusercontent.com/plotly/datasets/master/uber-rides-data1.csv"
 ridesRaw_1 <-  url1 %>%
   data.table::fread(stringsAsFactors = FALSE)
 url2 <- "https://raw.githubusercontent.com/plotly/datasets/master/uber-rides-data2.csv"
 ridesRaw_2 <-  url2 %>%
   data.table::fread(stringsAsFactors = FALSE)
 url3 <- "https://raw.githubusercontent.com/plotly/datasets/master/uber-rides-data3.csv"
 ridesRaw_3 <-  url3 %>%
   data.table::fread(stringsAsFactors = FALSE)
 ridesDf <- list(ridesRaw_1, ridesRaw_2, ridesRaw_3) %>%
   data.table::rbindlist()

 ridesDf %>%
       rasterly(mapping = aes(x = Lat, y = Lon)) %>%
       rasterly_points()
 }}