Static rasterly graphics

Render a rasterly object. This object can involve axes, legend, etc to better convert graphics to data.

rasterlyGrob(
  rasterlyObj,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  sub = NULL,
  interpolate = FALSE,
  axes = TRUE,
  legend = TRUE,
  legendlabel = NULL,
  legendlayer = 1,
  legendmain = NULL,
  axes_gpar = grid::gpar(col = "black", cex = 1),
  label_gpar = grid::gpar(col = "black", cex = 1),
  main_gpar = grid::gpar(col = "black", cex = 1.5),
  legend_gpar = grid::gpar(col = "black", cex = 1.5),
  name = NULL,
  gp = NULL,
  vp = NULL
)

grid.rasterly(
  rasterlyObj,
  interpolate = FALSE,
  axes = TRUE,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  sub = NULL,
  legend = TRUE,
  legendlabel = NULL,
  legendlayer = 1,
  legendmain = NULL,
  axes_gpar = grid::gpar(col = "black", cex = 1),
  label_gpar = grid::gpar(col = "black", cex = 1),
  main_gpar = grid::gpar(col = "black", cex = 1.5),
  legend_gpar = grid::gpar(col = "black", cex = 1.5),
  name = NULL,
  gp = NULL,
  vp = NULL,
  ...
)

# S3 method for rasterly
plot(
  x,
  y = NULL,
  xlim = NULL,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  legendmain = NULL,
  sub = NULL,
  interpolate = FALSE,
  axes = TRUE,
  legend = TRUE,
  legendlabel = NULL,
  legendlayer = 1,
  new.page = TRUE,
  ...
)

# S3 method for rasterly
print(x, ...)

Arguments

rasterlyObj
A rasterly object.

xlim
The x limits (x1, x2) of the plot.

ylim
The y limits (x1, x2) of the plot.

xlab
A y label for the plot.

ylab
A x label for the plot.

main
A title for the plot.

sub
A sub title for the plot.

interpolate
A logical value indicating whether to linearly interpolate the image (the alternative is to use nearest-neighbour interpolation, which gives a more blocky result).

axes
Show axes or not.

legend
Show legend or not.

legendlabel
The label of legend.

legendlayer
The legend represents which layer in rasterly object. Default is 1.

legendmain
The main title of legend

axes_gpar
gpar of axes. It is used to modify the axes' color, size and other aesthetics attributes.

label_gpar
gpar of label. It is used to modify the labels' color, size and other aesthetics attributes.

main_gpar
gpar of main. It is used to modify the main title's color, size and other aesthetics attributes.

legend_gpar
gpar of legend It is used to modify the main title's color, size and other aesthetics attributes.

name
a character identifier for the grob. Used to find the grob on the display list and/or as a child of another grob.

gp
A gpar object, typically the output from a call to the function gpar. This is basically a list of graphical parameter settings.

vp
a viewport object (or NULL).

...
Other arguments to modify the display.

x
A rasterly object

y
NULL, will be ignored.

new.page
display on a new page or not.

Details

We provide three functions to produce static graphics, which is based on the API of `grid`, `plot` and `print`.

grid: The rasterlyGrob and grid.rasterly are the most flexible data structure. It is nothing but a **grob**. Users can modify the existing display by the functions provided by grid.

plot.rasterly: It is a S3 method. The usage is very similar to the classic `plot` function. Users can set the x limits, y limits, x label, y label and etc.

print.rasterly: It is a S3 method, however, it only provides the most basic image raster.